ai-image-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
pydantic==2.6.4
python-multipart==0.0.9
numpy==1.26.4
pytest==8.1.1
httpx==0.27.0
